initramfs-framework: add support for upgrading to composefs
Add support for upgrading an OSTree system based on hard links (legacy)
into one based on composefs. The solution being introduced here performs
all its work within the initial ramdisk which means it adds quite a bit
of time to the boot process upon the upgrade. Though this is a one-time
process, it's still not ideal. In the future we may consider modifying
libostree to perform similar operation when generating a deployment
which would avoid the system downtime.

Related-to: TOR-3380

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>