am62xx: refactor the machine includes and factor out common options
Now that we have 3 AM62 Ti machines, some of the configurations we need
to change are common between then. To avoid duplicating everything,
those  machine includes now include a higher level 'am62xx.inc' file,
which has the common configurations for Torizon that are needed on the
platform.

Also now, Common Torizon distro, by default, will set the hostname as
the machine name we're building.

Related-to: TOR-3729

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>